# 🚀 Critical Phase 1 Implementation - Enhanced MCP Server with AI Router