--- a/Dockerfile.enhanced-mcp
+++ b/Dockerfile.enhanced-mcp
@@ -1,0 +1,46 @@
+# Dockerfile for Sophia Intel Enhanced MCP Server
+FROM python:3.11-slim
+
+# Set working directory
+WORKDIR /app
+
+# Install system dependencies
+RUN apt-get update && apt-get install -y \
+    curl \
+    git \
+    && rm -rf /var/lib/apt/lists/*
+
+# Copy requirements first for better caching
+COPY requirements.txt .
+
+# Install Python dependencies
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Install additional dependencies for enhanced server
+RUN pip install --no-cache-dir \
+    loguru \
+    aiohttp \
+    numpy \
+    python-dotenv
+
+# Copy application code
+COPY . .
+
+# Create non-root user for security
+RUN useradd -m -u 1000 sophia && chown -R sophia:sophia /app
+USER sophia
+
+# Expose port
+EXPOSE 8001
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
+    CMD curl -f http://localhost:8001/health || exit 1
+
+# Set environment variables
+ENV PYTHONPATH=/app
+ENV ENVIRONMENT=production
+
+# Run the simplified test server for now
+CMD ["python", "test_server_simple.py"]
+