--- a/pulumi/.gitignore
+++ b/pulumi/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/node_modules/