--- a/pulumi/index.ts
+++ b/pulumi/index.ts
@@ -1,0 +1,1 @@
+import * as pulumi from "@pulumi/pulumi";